refactor(ecma/parser): rm token contexts
